% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrcode.R
\name{get_qrcode_text}
\alias{get_qrcode_text}
\alias{qrcode}
\title{Get link of QR Code on SEFAZ-RS receipt}
\usage{
get_qrcode_text(url)
}
\arguments{
\item{image}{A image file of QR Code in SEFAZ-RS receipt.}
}
\description{
Get text from QR Code on SEFAZ-RS receipt.
}
