% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receipt.R
\name{get_sefaz_rs_rdata}
\alias{get_sefaz_rs_rdata}
\alias{receipt}
\title{Get receipt data from link of SEFAZ-RS}
\usage{
get_sefaz_rs_rdata(url)
}
\arguments{
\item{url}{A receipt url of https://www.sefaz.rs.gov.br/}
}
\description{
Get data from receipt in https://www.sefaz.rs.gov.br/
}
