% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receipt.R
\name{xml_table}
\alias{xml_table}
\title{Parse an html table into a data frame}
\usage{
xml_table(x, header = NA, trim = TRUE, fill = FALSE, dec = ".")
}
\arguments{
\item{x}{A node, node set or document.}

\item{header}{Use first row as header? If NA, will use first row if it consists of th tags.}

\item{trim}{Remove leading and trailing whitespace within each cell?}

\item{fill}{If TRUE, automatically fill rows with fewer than the maximum number of columns with NAs.}

\item{dec}{The character used as decimal mark.}
}
\description{
Parse an html table into a data frame
}
